+++ aptos-move/e2e-move-tests/src/tests/mod.rs
@@ -65,6 +65,7 @@ mod public_const_upgrade;
 mod public_struct_args;
 mod public_structs_enums_upgrade;
 mod randomness_test_and_abort;
+mod read_side_layout_dag;
 mod remote_state;
 mod resource_groups;
 mod rotate_auth_key;