+++ storage/aptosdb/src/pruner/state_merkle_pruner/test.rs
@@ -6,7 +6,9 @@ use crate::{
         test_helper::{arb_state_kv_sets_with_genesis, update_store},
         AptosDB,
     },
-    pruner::{PrunerManager, StateKvPrunerManager, StateMerklePrunerManager},
+    pruner::{
+        ColdStateKv, PrunerManager, StateKvPrunerManager, StateMerkle, StateMerklePrunerManager,
+    },
     schema::{
         stale_node_index::StaleNodeIndexSchema,
         stale_state_value_index_by_key_hash::StaleStateValueIndexByKeyHashSchema,
@@ -53,12 +55,15 @@ fn verify_state_in_store(
 fn create_state_merkle_pruner_manager(
     state_merkle_db: &Arc<StateMerkleDb>,
     prune_batch_size: usize,
-) -> StateMerklePrunerManager<StaleNodeIndexSchema> {
-    StateMerklePrunerManager::new(Arc::clone(state_merkle_db), StateMerklePrunerConfig {
-        enable: true,
-        prune_window: 0,
-        batch_size: prune_batch_size,
-    })
+) -> StateMerklePrunerManager<StateMerkle> {
+    StateMerklePrunerManager::<StateMerkle>::new(
+        Arc::clone(state_merkle_db),
+        StateMerklePrunerConfig {
+            enable: true,
+            prune_window: 0,
+            batch_size: prune_batch_size,
+        },
+    )
 }
 
 #[test]
@@ -316,12 +321,15 @@ fn verify_state_value_pruner(inputs: Vec<Vec<(StateKey, Option<StateValue>)>>) {
 
     let mut version = 0;
     let mut current_state_values = HashMap::new();
-    let pruner = StateKvPrunerManager::new(Arc::clone(&db.state_kv_db), LedgerPrunerConfig {
-        enable: true,
-        prune_window: 0,
-        batch_size: 1,
-        user_pruning_window_offset: 0,
-    });
+    let pruner = StateKvPrunerManager::<ColdStateKv>::new(
+        Arc::clone(&db.state_kv_db),
+        LedgerPrunerConfig {
+            enable: true,
+            prune_window: 0,
+            batch_size: 1,
+            user_pruning_window_offset: 0,
+        },
+    );
     for batch in inputs {
         update_store(store, batch.clone().into_iter(), version);
         for (k, v) in batch.iter() {